# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Creon API
  version: 0.1.0
  description: |
    Backend API for **Creon** — a Web3 crowdfunding (urun dana) platform for
    Indonesian UMKM on Stellar/Soroban.

    ## Response envelope
    Every JSON response is wrapped in a consistent envelope. Success:
    `{ statusCode, message, data }` — `data` holds exactly the resource shape
    documented per endpoint below; `message` is a short, human-readable string
    not meant for client-side branching (branch on `statusCode` / `data`
    instead). Errors: `{ statusCode, message, error, data: null }` — same
    `message`/`error` as before, now always paired with `data: null`. The one
    exception is `204 No Content` (`POST /auth/logout`), which has a
    genuinely empty body — no envelope.

    ## Authentication
    `POST /auth/login` or `POST /auth/register` set the JWT as an **httpOnly
    cookie** (`creon_access_token`) — it is not returned in the response body.
    Clients must send credentialed requests (`credentials: 'include'` for
    `fetch`, `withCredentials: true` for axios) so the cookie is attached
    automatically. Non-browser clients that cannot store cookies should
    re-send the cookie value via a `Cookie: creon_access_token=<jwt>` header.
    Bearer tokens are not accepted. `POST /auth/logout` clears the cookie.

    ### Wallet challenge-response flow
    1. `POST /auth/challenge` → receive a nonce message
    2. Sign the message with your Stellar secret key using SEP-53
...
servers:
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: creon_access_token
  schemas:
...
    Role:
      type: string
      enum: [ ENTREPRENEUR, INVESTOR, ADMIN ]
...
    ChallengeRequest:
      type: object
      required: [ walletAddress ]
      properties:
        walletAddress:
          type: string
          minLength: 56
          maxLength: 56
          description: Stellar public key (G…, 56 characters)
          example: GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV3MV72CSOFA
    RegisterRequest:
      type: object
      required: [ walletAddress, signature, role ]
      properties:
        walletAddress:
          type: string
          minLength: 56
          maxLength: 56
          description: Stellar public key (G…, 56 characters)
          example: GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV3MV72CSOFA
        signature:
          type: string
          description: Base64-encoded SEP-53 Ed25519 signature over the challenge message
        role:
          type: string
          enum: [ ENTREPRENEUR, INVESTOR ]
          description: Self-assignable role for the new account
        email:
          type: string
          format: email
          description: Required when `role` is `ENTREPRENEUR`; optional for investors
          example: founder@umkm.id
        displayName:
          type: string
          description: Optional display name
          example: Budi Santoso
    LoginRequest:
      type: object
      required: [ walletAddress, signature ]
      properties:
        walletAddress:
          type: string
          minLength: 56
          maxLength: 56
          example: GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV3MV72CSOFA
        signature:
          type: string
          description: Base64-encoded SEP-53 Ed25519 signature over the challenge message
...
    ChallengeResponse:
      type: object
      required: [ message ]
      properties:
        message:
          type: string
          description: Nonce string to sign with the Stellar secret key
          example: "Sign this message to authenticate with Creon: a3f8c1d2e9b7"
    AuthPrincipalResponse:
      type: object
      required: [ userId, roles ]
      description: |
        Returned by register/login after the JWT is set as an httpOnly cookie
        (`creon_access_token`) — the token itself is never in the body.
      properties:
        userId:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    UserProfileResponse:
      type: object
      required: [ id, walletAddress, roles, isActive, createdAt, updatedAt ]
      description: |
        Full user profile returned by `GET /auth/me`. Unlike `AuthPrincipalResponse`
        (which only contains `userId` and `roles`), this includes all profile fields.
      properties:
        id:
          type: string
          format: uuid
        walletAddress:
          type: string
          minLength: 56
          maxLength: 56
          example: GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV3MV72CSOFA
        email:
          type: [ string, 'null' ]
          format: email
          example: founder@umkm.id
        displayName:
          type: [ string, 'null' ]
          example: Budi Santoso
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [ statusCode, message, data ]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: "Both idCard and selfie files are required"
        error:
          type: string
          example: Bad Request
        data:
          type: 'null'
          example: null
    Envelope:
      type: object
      description: |
        Shared success envelope every JSON response is wrapped in (except
        `204 No Content`, which has a genuinely empty body). `data` is
        overridden per-endpoint below to the actual payload shape.
      required: [ statusCode, message, data ]
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data: {}
  responses:
    Unauthorized:
      description: Missing or invalid JWT bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not Found

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
  /auth/challenge:
    post:
      tags: [ auth ]
      summary: Dapatkan Pesan Challenge (Nonce)
      description: |
        Mengembalikan pesan nonce sekali pakai untuk ditandatangani oleh secret key Stellar user.

        **Untuk Frontend:**
        1. Panggil endpoint ini untuk mendapatkan pesan text.
        2. Minta user (lewat wallet/Freighter) untuk menandatangani pesan tersebut.
        3. Kirim hasil tanda tangan ke `/auth/login` atau `/auth/register`.

        *Catatan: Challenge ini akan kedaluwarsa dalam 5 menit.*
      operationId: authChallenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeRequest'
      responses:
        '200':
          description: Challenge message to sign
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChallengeResponse'
        '400':
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags: [ auth ]
      summary: Registrasi User Baru
      description: |
        Mendaftarkan akun user baru (Investor atau Entrepreneur) menggunakan hasil tanda tangan dari challenge.

        **Untuk Frontend:**
        - Pastikan sudah memanggil `/auth/challenge` sebelumnya.
        - Tanda tangani message menggunakan SEP-53, bukan raw bytes message.
        - Kirim role (`ENTREPRENEUR` atau `INVESTOR`).
        - Jika role `ENTREPRENEUR`, wajib mengirimkan `email`.
        - Jika berhasil, token JWT diset sebagai httpOnly cookie
          (`creon_access_token`) — body hanya berisi `userId` dan `roles`.
          Gunakan `credentials: 'include'` (fetch) / `withCredentials: true`
          (axios) agar cookie ikut terkirim di request berikutnya.
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created — JWT set as an httpOnly cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthPrincipalResponse'
        '400':
          description: Validation error or invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Wallet address already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [ auth ]
      summary: Login User
      description: |
        Autentikasi user yang sudah terdaftar menggunakan hasil tanda tangan dari challenge.

        **Untuk Frontend:**
        - Pastikan sudah memanggil `/auth/challenge` sebelumnya.
        - Kirim hasil tanda tangan SEP-53 ke endpoint ini.
        - Jika berhasil, token JWT diset sebagai httpOnly cookie
          (`creon_access_token`) — body hanya berisi `userId` dan `roles`.
          Gunakan `credentials: 'include'` (fetch) / `withCredentials: true`
          (axios) agar cookie ikut terkirim di request berikutnya.
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated — JWT set as an httpOnly cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthPrincipalResponse'
        '400':
          description: Invalid signature or expired challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Wallet address not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [ auth ]
      summary: Profil User Saat Ini
      description: |
        Mengembalikan data profil lengkap dari user yang sedang login
        berdasarkan JWT di httpOnly cookie.

        **Untuk Frontend:**
        - Panggil endpoint ini setelah login/register untuk mendapatkan
          data profil lengkap (wallet address, email, display name, roles, dll).
        - Gunakan `credentials: 'include'` (fetch) / `withCredentials: true`
          (axios) agar cookie JWT ikut terkirim.
        - Jika cookie tidak ada atau JWT sudah kedaluwarsa, response akan 401.
      operationId: authMe
      security:
        - CookieAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...